Refactor Paint Window into standalone widget
